# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart ERP & BI System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: assignedToMe
          in: query
          schema:
            type: boolean
        - name: status
          in: query
          schema:
            type: string
            enum: [new, in_progress, completed, cancelled]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, urgent]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/comments:
    post:
      operationId: addTaskComment
      tags: [tasks]
      summary: Add comment to task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentRequest"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskComment"
  /sales-targets:
    get:
      operationId: listSalesTargets
      tags: [tasks]
      summary: List sales targets
      responses:
        "200":
          description: Sales targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesTarget"
    post:
      operationId: createSalesTarget
      tags: [tasks]
      summary: Create sales target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesTargetRequest"
      responses:
        "201":
          description: Target created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesTarget"
  /sales-targets/{id}:
    delete:
      operationId: deleteSalesTarget
      tags: [tasks]
      summary: Delete sales target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Target deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, procurement, sales, inventory]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, name, email, role, isActive, createdAt]
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, procurement, sales, inventory]
      required: [name, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, procurement, sales, inventory]
        isActive:
          type: boolean
        password:
          type: string
...
    Task:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        createdById:
          type: integer
          nullable: true
        createdByName:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, urgent]
        status:
          type: string
          enum: [new, in_progress, completed, cancelled]
        dueDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, title, priority, status, createdAt]
    TaskDetail:
      allOf:
        - $ref: "#/components/schemas/Task"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/TaskComment"
          required: [comments]
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [low, medium, urgent]
        dueDate:
          type: string
          format: date
          nullable: true
      required: [title, priority]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, urgent]
        assignedToId:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
    TaskComment:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
          nullable: true
        userName:
          type: string
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, taskId, content, createdAt]
    AddCommentRequest:
      type: object
      properties:
        content:
          type: string
      required: [content]
    SalesTarget:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        currentStock:
          type: number
        targetQuantity:
          type: number
        soldQuantity:
          type: number
        progressPercent:
          type: number
        period:
          type: string
          enum: [week, month]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, productId, productName, productCode, currentStock, targetQuantity, soldQuantity, progressPercent, period, startDate, endDate, createdAt]
    CreateSalesTargetRequest:
      type: object
      properties:
        productId:
          type: integer
        targetQuantity:
          type: number
        period:
          type: string
          enum: [week, month]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required: [productId, targetQuantity, period, startDate, endDate]
...